--- src/Race.java
@@ -1,24 +1,0 @@
-import java.util.Random;
-
-public class Race {
-	
-	Random randomGenerator = new Random();
-	public String chosenRace;
-		
-     public Race() {
-    	 chosenRace= "";
-     }
-     
-     public void pickNewRace(){
-    	chosenRace= "";
-	   	int index = randomGenerator.nextInt(GenerateSourceData.getRaceSourceStatic().size());
-	   	chosenRace = GenerateSourceData.getRaceSourceStatic().get(index);
-	   	if(chosenRace.equals("Any Race")){
-	   		pickNewRace();
-	   	}
-     }
-     
-     public void pickNewRace(String race){
-    	 chosenRace = race;
-     }
-}
