--- src/data/SourceData.java
@@ -1,4 +1,0 @@
-package data;
-
-public class SourceData {
-}
