--- src/data/dtos/PersonalityList.java
@@ -1,17 +1,0 @@
-package data.dtos;
-
-import java.util.ArrayList;
-
-public class PersonalityList extends DataList {
-    public static PersonalityList of(ArrayList<String> list) {
-        return new PersonalityList(list);
-    }
-
-    public PersonalityList() {
-        this.list = new ArrayList<>();
-    }
-
-    private PersonalityList(ArrayList<String> list) {
-        this.list = list;
-    }
-}
