--- src/persistence/method/PersistenceMethod.java
@@ -1,8 +1,0 @@
-package persistence.method;
-
-import java.util.ArrayList;
-
-public interface PersistenceMethod {
-
-    ArrayList<String> GetData(String identifier);
-}
