+++ src/sanguinehaze/charactercreator/data/dtos/AdultProfessionList.java
@@ -1,0 +1,19 @@
+package sanguinehaze.charactercreator.data.dtos;
+
+import java.util.ArrayList;
+
+public class AdultProfessionList extends DataList {
+
+    public static AdultProfessionList of(ArrayList<String> list) {
+        return new AdultProfessionList(list);
+    }
+
+    public AdultProfessionList() {
+        this.list = new ArrayList<>();
+    }
+
+    private AdultProfessionList(ArrayList<String> list) {
+        this.list = list;
+    }
+
+}
