+++ src/sanguinehaze/charactercreator/persistence/SourcePersistanceFactory.java
@@ -1,0 +1,12 @@
+package sanguinehaze.charactercreator.persistence;
+
+import sanguinehaze.charactercreator.persistence.method.FileSystemPersistenceMethod;
+import sanguinehaze.charactercreator.persistence.method.PersistenceMethod;
+
+public class SourcePersistanceFactory {
+    public static SourcePersistence Create() {
+
+        PersistenceMethod persistanceMethod = new FileSystemPersistenceMethod();
+        return new SourcePersistence(persistanceMethod);
+    }
+}
