+++ src/sanguinehaze/charactercreator/persistence/method/PersistenceMethod.java
@@ -1,0 +1,8 @@
+package sanguinehaze.charactercreator.persistence.method;
+
+import java.util.ArrayList;
+
+public interface PersistenceMethod {
+
+    ArrayList<String> GetData(String identifier);
+}
